Custom tmpdir arg (first try)
Attempt to add a --tmpdir arg.

The --shared option can be very useful but also a little dangerous if
you don’t trust what’s in the /tmp directory. Giving Arx a custom
location for the temp directory should make things a little bit safer.
Most useful in this case,

  $ arx tmpx --shared --tmpdir '$HOME/.cache' my.tar.bz2

This will put things in your HOME directory where you can be fairly
sure no one else has snuck in a temporary directory to fool you into
using something that is malicious.

Using /tmp as a shared directory can be exploited if you know the hash
ahead of time. Just something as simple as this,

  $ tmp=$(mktemp -d tmpx-HASH)
  $ echo "rm -rf /" > $tmp/run

  where HASH is the hash of the dat,

can create a privilege escalation when Arx runs the /run script.